requirements.txt
pyrum==0.1.0
pytest==5.4.1
pytest-trio==0.5.2
git+https://github.com/ethereum/eth2.0-specs@v0.11.1
git+https://github.com/protolambda/eth2fastspec.git
